Was meant to create a separate branch for this, but the work ended up being an expansion of Descriptors.
I added and completed the createTextureView and createTextureSampler functions.
Also updated them in the InitVulkan and Cleanup functions.

Expanded Descriptors to include a combined sampler for the texture.

Updated both the Vertex Structs member variables and functions to accomodate textures, as well as the Vertex and Fragment shaders.

Comments on all the stuff that was new.