this is an intermediary state commit.
I have made a struct called Mesh for storing vertices and indices, and have tested using it to replace the staticly declared vertices and indices.

I have also introduced the skeleton code for loading multiple Meshes, and the next step will be to introduce storing the Meshes in a map of sorts, so that we can use key association in conjunction with map searches when we need it.